+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Version History
 ---------------
 
+-   Added oidnGetBufferData and oidnGetBufferSize functions
+
 ### Changes in v1.4.1:
 
 -   Fixed crash when in-place denoising images with certain unusual resolutions